chore: move select component to components dir
